+++ packages/ui/src/index.ts
@@ -188,6 +188,7 @@ export {
   type TrackShopifyProps,
 } from './Track/index.js';
 export { Transition, type TransitionConstructor } from './Transition/index.js';
+export type { DomUpdateRunner, DomUpdateTransitioner } from './utils/dom-update.js';
 export {
   viewTransition,
   ViewTransition,